+++ CHANGELOG.md
@@ -23,6 +23,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - [Move transplant_days field to farm_transplant module #795](https://github.com/farmOS/farmOS/pull/795)
 - [Use content form for taxonomy terms #810](https://github.com/farmOS/farmOS/pull/810)
 - [Override the lab_test log timestamp label and description #774](https://github.com/farmOS/farmOS/pull/774)
+- [Override the birth log asset label and description #824](https://github.com/farmOS/farmOS/pull/824)
 
 ### Fixed